pyproject.toml
[tool.poetry]
name = "voice-assistant"
version = "0.1.0"
description = ""
authors = ["Arunachalam <m10arunachalam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
django-cors-headers = "4.3.1"
azure-cognitiveservices-speech = "1.34.0"
sentence-transformers = "2.7.0"
openai = "^1.10.0"
pyttsx3 = "2.90"
chromadb = "0.4.24"
streamlit = "1.28.0"
requests = "2.31.0"
python-dotenv = "1.0.0"
pypdf2 = "3.0.1"
huggingface-hub = ">=0.15.0,<0.20.0"
audio-recorder-streamlit = "^0.0.10"
httpx = "0.27.2"
langchain = "^0.1.0"
langchain-community = "^0.0.20"
langchain-openai = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
